issue #5 work

+++ source/m3_core.h
@@ -48,7 +48,7 @@ typedef const u8 *              bytes_t;
 typedef const u8 * const        cbytes_t;
 
 typedef i64                     m3reg_t;
-typedef u64                     m3slot_t;
+typedef u32                     m3slot_t;
 typedef m3slot_t *              m3stack_t;
 
 typedef
